+++ src/main/java/baseball/domain/gamenumbercreator/GameNumberCreator.java
@@ -1,5 +1,6 @@
-package baseball.domain;
+package baseball.domain.gamenumbercreator;
 
+import baseball.domain.GameNumber;
 import java.util.List;
 
 public interface GameNumberCreator {
