+++ src/test/java/baseball/domain/gamenumbercreator/StringToGameNumberCreatorTest.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.gamenumbercreator;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
